# temp_openapi_post_orch.json
{"openapi":"3.1.0","info":{"title":"PCD EsSalud â ANEXO 2","version":"0.1.0"},"paths":{"/reload":{"post":{"summary":"Reload Data","description":"Recarga el archivo JSON en `DATA_PATH` y actualiza la lista en memoria.\n\nÃtil en desarrollo para evitar reiniciar Uvicorn cuando los datos cambian.","operationId":"reload_data_reload_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/series":{"get":{"summary":"Listar Series","operationId":"listar_series_series_get","parameters":[{"name":"q","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Q"}},{"name":"fondo","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Fondo"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Serie"},"title":"Response Listar Series Series Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/series/{codigo}":{"get":{"summary":"Por Codigo","description":"Return a series by its code.\n\nAccepts codes that contain slashes (e.g. 'CODI/01') by using a path-style\nparameter so the client does not need to URL-encode the slash.","operationId":"por_codigo_series__codigo__get","parameters":[{"name":"codigo","in":"path","required":true,"schema":{"type":"string","title":"Codigo"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Serie"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/fondos":{"get":{"summary":"Fondos","operationId":"fondos_fondos_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"integer"},"type":"object","title":"Response Fondos Fondos Get"}}}}}}},"/trd/classify":{"post":{"summary":"Trd Classify","description":"Classify a document according to the TRD rules and return suggested code and disposition.","operationId":"trd_classify_trd_classify_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TRDRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Retencion":{"properties":{"gestion":{"type":"integer","title":"Gestion"},"periferico":{"type":"integer","title":"Periferico"},"central":{"type":"integer","title":"Central"},"total":{"type":"integer","title":"Total"}},"type":"object","required":["gestion","periferico","central","total"],"title":"Retencion"},"Serie":{"properties":{"fondo":{"type":"string","title":"Fondo"},"codigo":{"type":"string","title":"Codigo"},"titulo":{"type":"string","title":"Titulo"},"valor":{"type":"string","title":"Valor"},"retencion":{"$ref":"#/components/schemas/Retencion"},"observaciones":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Observaciones"}},"type":"object","required":["fondo","codigo","titulo","valor","retencion"],"title":"Serie"},"TRDRequest":{"properties":{"serie":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Serie"},"subserie":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Subserie"},"tipo":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Tipo"},"titulo":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Titulo"},"productor":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Productor"},"asuntoUnidad":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Asuntounidad"}},"type":"object","title":"TRDRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}